[BLE] move generic code from demo_central.rb to ble_central.rb
